pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataalchemy"
version = "0.1.0"
description = "DataAlchemy: a data-distribution lens on Chain-of-Thought reasoning."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1,<3",
    "transformers>=4.44,<5",
    "datasets>=2.19,<4",
    "accelerate>=0.30,<2",
    "tokenizers>=0.19,<1",
    "sentencepiece>=0.2,<1",
    "numpy>=1.26,<3",
    "pyyaml>=6.0,<7",
    "editdistance>=0.8,<1",
    "nltk>=3.8,<4",
    "tqdm>=4.66,<5",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.5"]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -m 'not slow'"
markers = [
    "slow: integration tests that train a real model (opt-in with `-m slow`)",
]
filterwarnings = [
    "ignore:builtin type Swig.* has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
    "ignore:Deprecated in 0\\.9\\.0:DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [".venv", "saves", "logs", "results", "data"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]
# Scripts intentionally manipulate sys.path before imports.
"scripts/*" = ["E402"]
